incorporating hdf5 into training routine, refactoring dataset creation for arrays
