+++ src/classpose/train.py
@@ -8,15 +8,13 @@ from typing import Any
 import numpy as np
 import torch
 import torch.distributed as dist
-from cellpose import dynamics, models, utils
-from cellpose.train import _loss_fn_seg, _reshape_norm
-from skimage import io
+from cellpose.train import _loss_fn_seg
 from torch import nn
 from torch.utils.data import DataLoader
-from tqdm import trange
 
 from classpose.dataset import (
     ClassposeTrainingDataset,
+    ClassposeHDF5Dataset,
     DistributedEpochSampler,
     SequentialDistributedSampler,
 )
@@ -49,7 +47,8 @@ class LossAggregator(nn.Module):
         """
         Args:
             n_losses (int): Number of losses to aggregate
-            optimise (bool): Whether to learn uncertainty weights (True) or use equal weighting (False)
+            optimise (bool): Whether to learn uncertainty weights (True) or use
+                equal weighting (False)
         """
         super().__init__()
         self.n_losses = n_losses
@@ -180,305 +179,6 @@ def _loss_fn_class(
     loss3 = criterion3(y[:, :-3], lbl[:, 0].long())
 
     return loss3
-
-
-def split_labels(
-    labels: list[np.ndarray], mask_classes: bool = True
-) -> tuple[list[np.ndarray], list[np.ndarray]]:
-    """
-    Split labels into class and mask arrays.
-
-    Args:
-        labels (list[np.ndarray]): List of label arrays.
-        mask_classes (bool, optional): masks class labels. Defaults to True.
-            Class labels are masked out if their corresponding foreground pixel
-            is not foreground.
-
-    Returns:
-        tuple[list[np.ndarray], list[np.ndarray]]: Tuple of class and mask arrays.
-    """
-    classes = [labels[i][-1:] for i in range(len(labels))]
-    labels = [labels[i][:-1] for i in range(len(labels))]
-    if mask_classes:
-        for i in range(len(classes)):
-            classes[i] = np.int16(classes[i])
-            # mask class labels where instance foreground is 0
-            classes[i][np.logical_and(labels[i][0] == 0, classes[i] > 0)] = -100
-            # mask class labels where class foreground is 0 but instance foreground is > 0
-            classes[i][np.logical_and(labels[i][0] > 0, classes[i] == 0)] = -100
-    return labels, classes
-
-
-def _process_train_test(
-    train_data: list[np.ndarray] | None = None,
-    train_labels: list[np.ndarray] | None = None,
-    train_files: list[str] | None = None,
-    train_labels_files: list[str] | None = None,
-    train_probs: np.ndarray | None = None,
-    train_classes: np.ndarray | None = None,
-    test_data: list[np.ndarray] | None = None,
-    test_labels: list[np.ndarray] | None = None,
-    test_files: list[str] | None = None,
-    test_labels_files: list[str] | None = None,
-    test_probs: np.ndarray | None = None,
-    test_classes: np.ndarray | None = None,
-    load_files: bool = True,
-    min_train_masks: int = 5,
-    compute_flows: bool = False,
-    normalize_params: dict = {"normalize": False},
-    channel_axis: int | None = None,
-    device: torch.device | None = None,
-):
-    """
-    Process train and test data.
-
-    The main change when compared with Cellpose is that, if not provided,
-    class labels are expected to be the last channel in the label array. As
-    such, the label array should have 2 (instance + class) or 4
-    (instance + flows (2) + class) channels.
-
-    Args:
-        train_data (list[np.ndarray] or None): List of training data arrays.
-        train_labels (list[np.ndarray] or None): List of training label arrays.
-        train_files (list[str] or None): List of training file paths.
-        train_labels_files (list[str] or None): List of training label file paths.
-        train_probs (np.ndarray or None): Array of training probabilities.
-        train_classes (np.ndarray or None): Array of training classes.
-        test_data (list[np.ndarray] or None): List of test data arrays.
-        test_labels (list[np.ndarray] or None): List of test label arrays.
-        test_files (list[str] or None): List of test file paths.
-        test_labels_files (list[str] or None): List of test label file paths.
-        test_probs (np.ndarray or None): Array of test probabilities.
-        test_classes (np.ndarray or None): Array of test classes.
-        load_files (bool): Whether to load data from files.
-        min_train_masks (int): Minimum number of masks required for training images.
-        compute_flows (bool): Whether to compute flows.
-        channels (list[int] or None): List of channel indices to use.
-        channel_axis (int or None): Axis of channel dimension.
-        rgb (bool): Convert training/testing images to RGB.
-        normalize_params (dict): Dictionary of normalization parameters.
-        device (torch.device): Device to use for computation.
-
-    Returns:
-        tuple: A tuple containing the processed train and test data and sampling probabilities and diameters.
-    """
-    device = get_default_device(device)
-
-    if train_data is not None and train_labels is not None:
-        # if data is loaded
-        nimg = len(train_data)
-        nimg_test = len(test_data) if test_data is not None else None
-    else:
-        # otherwise use files
-        nimg = len(train_files)
-        if train_labels_files is None:
-            train_labels_files = [
-                os.path.splitext(str(tf))[0] + "_flows.tif"
-                for tf in train_files
-            ]
-            train_labels_files = [
-                tf for tf in train_labels_files if os.path.exists(tf)
-            ]
-        if (
-            test_data is not None or test_files is not None
-        ) and test_labels_files is None:
-            test_labels_files = [
-                os.path.splitext(str(tf))[0] + "_flows.tif" for tf in test_files
-            ]
-            test_labels_files = [
-                tf for tf in test_labels_files if os.path.exists(tf)
-            ]
-        if not load_files:
-            train_logger.info(">>> using files instead of loading dataset")
-        else:
-            # load all images
-            train_logger.info(">>> loading images and labels")
-            train_data = [io.imread(train_files[i]) for i in trange(nimg)]
-            train_labels = [
-                io.imread(train_labels_files[i]) for i in trange(nimg)
-            ]
-        nimg_test = len(test_files) if test_files is not None else None
-        if load_files and nimg_test:
-            test_data = [io.imread(test_files[i]) for i in trange(nimg_test)]
-            test_labels = [
-                io.imread(test_labels_files[i]) for i in trange(nimg_test)
-            ]
-
-    ### check that arrays are correct size
-    if (train_labels is not None and nimg != len(train_labels)) or (
-        train_labels_files is not None and nimg != len(train_labels_files)
-    ):
-        error_message = "train data and labels not same length"
-        train_logger.critical(error_message)
-        raise ValueError(error_message)
-    if (test_labels is not None and nimg_test != len(test_labels)) or (
-        test_labels_files is not None and nimg_test != len(test_labels_files)
-    ):
-        train_logger.warning("test data and labels not same length, not using")
-        test_data, test_files = None, None
-    if train_labels is not None:
-        if train_labels[0].ndim < 2 or train_data[0].ndim < 2:
-            error_message = (
-                "training data or labels are not at least two-dimensional"
-            )
-            train_logger.critical(error_message)
-            raise ValueError(error_message)
-        if train_data[0].ndim > 3:
-            error_message = "training data is more than three-dimensional (should be 2D or 3D array)"
-            train_logger.critical(error_message)
-            raise ValueError(error_message)
-
-        # perform here classpose-specific checks
-        # we first check that labels have three dimensions
-        if train_labels[0].ndim == 3:
-            # check that the first channel is the instance label
-            if train_labels[0].shape[0] == 2 or train_labels[0].shape[0] == 4:
-                if train_classes is None:
-                    train_labels, train_classes = split_labels(train_labels)
-                else:
-                    train_labels, _ = split_labels(train_labels)
-                if test_labels is not None:
-                    if test_classes is None:
-                        test_labels, test_classes = split_labels(test_labels)
-                    else:
-                        test_labels, _ = split_labels(test_labels)
-            else:
-                error_message = "training labels should have two or four channels for classpose if classes are not provided"
-                train_logger.critical(error_message)
-                raise ValueError(error_message)
-        else:
-            error_message = "training labels should have three dimensions"
-            train_logger.critical(error_message)
-            raise ValueError(error_message)
-
-    ### check that flows are computed
-    if train_labels is not None:
-        train_labels = dynamics.labels_to_flows(
-            train_labels, files=train_files, device=device
-        )
-        if test_labels is not None:
-            test_labels = dynamics.labels_to_flows(
-                test_labels, files=test_files, device=device
-            )
-    elif compute_flows:
-        for k in trange(nimg):
-            tl = dynamics.labels_to_flows(
-                io.imread(train_labels_files[k]),
-                files=train_files,
-                device=device,
-            )
-        if test_files is not None:
-            for k in trange(nimg_test):
-                tl = dynamics.labels_to_flows(
-                    io.imread(test_labels_files[k]),
-                    files=test_files,
-                    device=device,
-                )
-
-    ### compute diameters
-    nmasks = np.zeros(nimg)
-    diam_train = np.zeros(nimg)
-    train_logger.info(">>> computing diameters")
-    for k in trange(nimg):
-        tl = (
-            train_labels[k][0]
-            if train_labels is not None
-            else io.imread(train_labels_files[k])[0]
-        )
-        diam_train[k], dall = utils.diameters(tl)
-        nmasks[k] = len(dall)
-    diam_train[diam_train < 5] = 5.0
-    if test_data is not None:
-        diam_test = np.array(
-            [
-                utils.diameters(test_labels[k][0])[0]
-                for k in trange(len(test_labels))
-            ]
-        )
-        diam_test[diam_test < 5] = 5.0
-    elif test_labels_files is not None:
-        diam_test = np.array(
-            [
-                utils.diameters(io.imread(test_labels_files[k])[0])[0]
-                for k in trange(len(test_labels_files))
-            ]
-        )
-        diam_test[diam_test < 5] = 5.0
-    else:
-        diam_test = None
-
-    ### check to remove training images with too few masks
-    if min_train_masks > 0:
-        nremove = (nmasks < min_train_masks).sum()
-        if nremove > 0:
-            train_logger.warning(
-                f"{nremove} train images with number of masks less than min_train_masks ({min_train_masks}), removing from train set"
-            )
-            ikeep = np.nonzero(nmasks >= min_train_masks)[0]
-            if train_data is not None:
-                train_data = [train_data[i] for i in ikeep]
-                train_labels = [train_labels[i] for i in ikeep]
-            if train_files is not None:
-                train_files = [train_files[i] for i in ikeep]
-            if train_labels_files is not None:
-                train_labels_files = [train_labels_files[i] for i in ikeep]
-            if train_probs is not None:
-                train_probs = train_probs[ikeep]
-            if train_classes is not None:
-                train_classes = [train_classes[i] for i in ikeep]
-            diam_train = diam_train[ikeep]
-            nimg = len(train_data)
-
-    ### normalize probabilities
-    train_probs = (
-        1.0 / nimg * np.ones(nimg, "float64")
-        if train_probs is None
-        else train_probs
-    )
-    train_probs /= train_probs.sum()
-    if test_files is not None or test_data is not None:
-        test_probs = (
-            1.0 / nimg_test * np.ones(nimg_test, "float64")
-            if test_probs is None
-            else test_probs
-        )
-        test_probs /= test_probs.sum()
-
-    ### reshape and normalize train / test data
-    normed = False
-    if normalize_params["normalize"]:
-        train_logger.info(f">>> normalizing {normalize_params}")
-    if train_data is not None:
-        train_data = _reshape_norm(
-            train_data,
-            channel_axis=channel_axis,
-            normalize_params=normalize_params,
-        )
-        normed = True
-    if test_data is not None:
-        test_data = _reshape_norm(
-            test_data,
-            channel_axis=channel_axis,
-            normalize_params=normalize_params,
-        )
-
-    return (
-        train_data,
-        train_labels,
-        train_files,
-        train_labels_files,
-        train_probs,
-        train_classes,
-        diam_train,
-        test_data,
-        test_labels,
-        test_files,
-        test_labels_files,
-        test_probs,
-        test_classes,
-        diam_test,
-        normed,
-    )
 
 
 def seed_everything(seed: int):
@@ -655,39 +355,21 @@ def _should_validate(iepoch: int, validate_every_epoch: bool) -> bool:
 
 def train_class_seg(
     net: torch.nn.Module,
-    train_data: list[np.ndarray] | None = None,
-    train_labels: list[np.ndarray] | None = None,
-    train_files: list[str] | None = None,
-    train_labels_files: list[str] | None = None,
-    train_probs: list[float] | None = None,
-    train_classes: list[int] | None = None,
-    test_data: list[np.ndarray] | None = None,
-    test_labels: list[np.ndarray] | None = None,
-    test_files: list[str] | None = None,
-    test_labels_files: list[str] | None = None,
-    test_probs: list[float] | None = None,
-    test_classes: list[int] | None = None,
-    channel_axis: int | None = None,
-    load_files: bool = True,
+    train_dataset: ClassposeHDF5Dataset | ClassposeTrainingDataset,
+    train_probs: np.ndarray | None = None,
+    test_dataset: ClassposeHDF5Dataset | ClassposeTrainingDataset | None = None,
     batch_size: int = 1,
     learning_rate: float | list[float] = 5e-5,
-    SGD: bool = False,
     n_epochs: int = 100,
     weight_decay: float = 0.1,
-    normalize: bool | dict = True,
-    compute_flows: bool = False,
     save_path: str | None = None,
     save_every: int = 100,
     save_each: bool = False,
     nimg_per_epoch: int | None = None,
     nimg_test_per_epoch: int | None = None,
-    rescale: bool = False,
     scale_range: list[float] | None = None,
-    bsize: int = 256,
-    min_train_masks: int = 5,
     model_name: str | None = None,
     class_weights: list[float] | None = None,
-    augmentation_strategy: str = "hed_only",
     num_workers: int = 4,
     use_uncertainty_weighting: bool = False,
     validate_every_epoch: bool = False,
@@ -705,41 +387,25 @@ def train_class_seg(
 
     Args:
         net (torch.nn.Module): The network model to train.
-        train_data (list[np.ndarray], optional): List of arrays (2D or 3D) - images for training. Defaults to None.
-        train_labels (list[np.ndarray], optional): List of arrays (2D or 3D) - labels for train_data, where 0=no masks; 1,2,...=mask labels. Defaults to None.
-        train_files (list[str], optional): List of strings - file names for images in train_data (to save flows for future runs). Defaults to None.
-        train_labels_files (list, optional): List of training label file paths. Defaults to None.
-        train_probs (list[float], optional): List of floats - probabilities for each image to be selected during training. Defaults to None.
-        train_classes (list[int], optional): Per-instance class labels aligned with train_labels when they are provided separately. Defaults to None.
-        test_data (list[np.ndarray], optional): List of arrays (2D or 3D) - images for testing. Defaults to None.
-        test_labels (list[np.ndarray], optional): List of arrays (2D or 3D) - labels for test_data, where 0=no masks; 1,2,...=mask labels. Defaults to None.
-        test_files (list[str], optional): List of strings - file names for images in test_data (to save flows for future runs). Defaults to None.
-        test_labels_files (list, optional): List of test label file paths. Defaults to None.
-        test_probs (list[float], optional): List of floats - probabilities for each image to be selected during testing. Defaults to None.
-        test_classes (list[int], optional): Per-instance class labels aligned with test_labels when they are provided separately. Defaults to None.
-        channel_axis (int | None, optional): Channel dimension for raw image arrays before reshaping/normalization. Defaults to None.
-        load_files (bool, optional): Boolean - whether to load images and labels from files. Defaults to True.
+        train_dataset (ClassposeHDF5Dataset | ClassposeTrainingDataset): Dataset containing training images and labels.
+        train_probs (np.ndarray | None, optional): Probabilities for each image to be selected during training. Defaults to None.
+        test_dataset (ClassposeHDF5Dataset | ClassposeTrainingDataset | None, optional): Dataset containing testing images and labels. Defaults to None.
         batch_size (int, optional): Number of training samples processed per optimizer step on the current process. Defaults to 1.
         learning_rate (float | list[float], optional): Base learning rate used to build the warmup/decay schedule. Defaults to 5e-5.
         n_epochs (int, optional): Number of training epochs to run. Defaults to 100.
         weight_decay (float, optional): Weight decay for AdamW. Defaults to 0.1.
-        SGD (bool, optional): Deprecated in v4.0.1+ - AdamW always used.
-        normalize (bool | dict, optional): Boolean or dictionary - whether to normalize the data. Defaults to True.
-        compute_flows (bool, optional): Boolean - whether to compute flows during training. Defaults to False.
-        save_path (str, optional): String - where to save the trained model. Defaults to None.
+        save_path (str | None, optional): String - where to save the trained model. Defaults to None.
         save_every (int, optional): Integer - save the network every [save_every] epochs. Defaults to 100.
         save_each (bool, optional): Boolean - save the network to a new filename at every [save_each] epoch. Defaults to False.
-        nimg_per_epoch (int, optional): Integer - minimum number of images to train on per epoch. Defaults to None.
-        nimg_test_per_epoch (int, optional): Integer - minimum number of images to test on per epoch. Defaults to None.
-        rescale (bool, optional): Whether or not to rescale images during training. Defaults to False.
-        min_train_masks (int, optional): Integer - minimum number of masks an image must have to use in the training set. Defaults to 5.
-        model_name (str, optional): String - name of the network. Defaults to None.
-        class_weights (list[float], optional): List of class weights for weighted loss computation. Defaults to None.
-        augmentation_strategy (str, optional): Pre-defined augmentation strategy name. Options: 'hed_only', 'enhanced'. Defaults to 'hed_only'.
+        nimg_per_epoch (int | None, optional): Integer - minimum number of images to train on per epoch. Defaults to None.
+        nimg_test_per_epoch (int | None, optional): Integer - minimum number of images to test on per epoch. Defaults to None.
+        scale_range (list[float] | None, optional): Range of scales for data augmentation. Defaults to None.
+        model_name (str | None, optional): String - name of the network. Defaults to None.
+        class_weights (list[float] | None, optional): List of class weights for weighted loss computation. Defaults to None.
         num_workers (int, optional): Number of DataLoader workers used for data loading. Defaults to 4.
         use_uncertainty_weighting (bool, optional): Whether to use task uncertainty weighting for automatic loss balancing. If True, the model learns optimal weights for segmentation, classification CE, and Tversky losses automatically. If False, uses equal weighting (1.0 each). Defaults to False.
         validate_every_epoch (bool, optional): If True, run validation at every epoch instead of the legacy sparse schedule. Defaults to False.
-        log_file_path (str, optional): Path to the log file. Defaults to None.
+        log_file_path (str | None, optional): Path to the log file. Defaults to None.
         random_seed (int, optional): Base random seed. The effective seed is offset by rank in distributed training. Defaults to 42.
         device (torch.device | None, optional): Explicit device for training and checkpoint restore. If None, it is derived from the model/device helpers. Defaults to None.
         distributed (bool | None, optional): Whether distributed training is active. If None, the function infers it from torch.distributed state. Defaults to None.
@@ -758,9 +424,6 @@ def train_class_seg(
     if log_file_path is not None and is_main_process():
         add_file_handler(train_logger, log_file_path)
 
-    if SGD:
-        train_logger.warning("SGD is deprecated, using AdamW instead")
-
     raw_net = unwrap_model(net)
     if device is None:
         try:
@@ -772,64 +435,11 @@ def train_class_seg(
 
     scale_range = 0.5 if scale_range is None else scale_range
 
-    if isinstance(normalize, dict):
-        normalize_params = {**models.normalize_default, **normalize}
-    elif not isinstance(normalize, bool):
-        raise ValueError("normalize parameter must be a bool or a dict")
-    else:
-        normalize_params = models.normalize_default.copy()
-        normalize_params["normalize"] = normalize
-
     oversampling_active = train_probs is not None
 
-    out = _process_train_test(
-        train_data=train_data,
-        train_labels=train_labels,
-        train_files=train_files,
-        train_labels_files=train_labels_files,
-        train_probs=train_probs,
-        train_classes=train_classes,
-        test_data=test_data,
-        test_labels=test_labels,
-        test_files=test_files,
-        test_labels_files=test_labels_files,
-        test_probs=test_probs,
-        test_classes=test_classes,
-        load_files=load_files,
-        min_train_masks=min_train_masks,
-        compute_flows=compute_flows,
-        channel_axis=channel_axis,
-        normalize_params={"normalize": False},
-        device=device,
+    raw_net.diam_labels.data = torch.tensor(
+        [train_dataset.diameter_array.mean()], device=device
     )
-    (
-        train_data,
-        train_labels,
-        train_files,
-        train_labels_files,
-        train_probs,
-        train_classes,
-        diam_train,
-        test_data,
-        test_labels,
-        test_files,
-        test_labels_files,
-        test_probs,
-        test_classes,
-        diam_test,
-        _normed,
-    ) = out
-    train_labels = [
-        np.concatenate((y[:1], y_class, y[1:]), axis=0)
-        for y, y_class in zip(train_labels, train_classes)
-    ]
-    if test_labels is not None:
-        test_labels = [
-            np.concatenate((y[:1], y_class, y[1:]), axis=0)
-            for y, y_class in zip(test_labels, test_classes)
-        ]
-
-    raw_net.diam_labels.data = torch.tensor([diam_train.mean()], device=device)
 
     if class_weights is not None and isinstance(
         class_weights, (list, np.ndarray, tuple)
@@ -837,9 +447,8 @@ def train_class_seg(
         class_weights = np.float32(class_weights)
         class_weights = torch.from_numpy(class_weights).to(device).float()
 
-    nimg = len(train_data) if train_data is not None else len(train_files)
-    nimg_test = len(test_data) if test_data is not None else None
-    nimg_test = len(test_files) if test_files is not None else nimg_test
+    nimg = len(train_dataset)
+    nimg_test = len(train_dataset)
     nimg_per_epoch = nimg if nimg_per_epoch is None else nimg_per_epoch
     if nimg_test is not None and nimg_test_per_epoch not in (None, nimg_test):
         train_logger.warning(
@@ -913,30 +522,13 @@ def train_class_seg(
     train_losses, test_losses = np.zeros(n_epochs), np.zeros(n_epochs)
     best_val_loss = np.inf
 
-    train_logger.info(
-        f">>> Using augmentation strategy: '{augmentation_strategy}'"
-    )
     if num_workers > 0:
         train_logger.info(
-            f">>> Using multiprocessing for training with {num_workers} workers"
+            f"Using multiprocessing for training with {num_workers} workers"
         )
     else:
-        train_logger.info(
-            ">>> Using single-threaded processing for training"
-        )
+        train_logger.info("Using single-threaded processing for training")
 
-    train_dataset = ClassposeTrainingDataset(
-        data_array=train_data,
-        label_array=train_labels,
-        diameter_array=diam_train,
-        diam_mean=raw_net.diam_mean.item(),
-        rescale=rescale,
-        scale_range=scale_range,
-        bsize=bsize,
-        normalize_params=normalize_params,
-        augment=True,
-        augment_pipeline_config=augmentation_strategy,
-    )
     train_sampler = DistributedEpochSampler(
         dataset_length=nimg,
         train_probs=train_probs if oversampling_active else None,
@@ -958,19 +550,7 @@ def train_class_seg(
     )
 
     val_loader = None
-    if test_data is not None or test_files is not None:
-        test_dataset = ClassposeTrainingDataset(
-            data_array=test_data,
-            label_array=test_labels,
-            diameter_array=diam_test,
-            diam_mean=raw_net.diam_mean.item(),
-            rescale=rescale,
-            scale_range=scale_range,
-            bsize=bsize,
-            normalize_params=normalize_params,
-            augment=True,
-            augment_pipeline_config=None,
-        )
+    if test_dataset is not None:
         val_sampler = (
             SequentialDistributedSampler(
                 dataset_length=nimg_test,
@@ -1149,7 +729,10 @@ def train_class_seg(
                     f"LR={optimizer.param_groups[0]['lr']:.6f}, time {time.time()-t0:.2f}s"
                 )
 
-        if _should_validate(iepoch, validate_every_epoch) and val_loader is not None:
+        if (
+            _should_validate(iepoch, validate_every_epoch)
+            and val_loader is not None
+        ):
             net.eval()
             val_seg_sum = 0.0
             val_ce_sum = 0.0
@@ -1191,7 +774,9 @@ def train_class_seg(
                     val_sample_count += batch_size_local
                     val_seg_sum += batch_val_seg_loss * batch_size_local
                     val_ce_sum += loss_ce_val.item() * batch_size_local
-                    val_tversky_sum += loss_tversky_val.item() * batch_size_local
+                    val_tversky_sum += (
+                        loss_tversky_val.item() * batch_size_local
+                    )
                     val_total_sum += loss.item() * batch_size_local
 
             val_totals = torch.tensor(
