+++ src/classpose/utils.py
@@ -264,45 +264,6 @@ def oversample_classes(
     return X, Y
 
 
-def get_class_counts(Y: np.ndarray, n_classes: int) -> np.ndarray:
-    """
-    Get class counts.
-
-    Args:
-        Y (np.ndarray): Array of labels (2 channels: instance and class).
-        n_classes (int): Number of classes.
-
-    Returns:
-        np.ndarray: Array of class counts.
-    """
-    return np.bincount(Y, minlength=n_classes)
-
-
-def get_class_weights(
-    Y: np.ndarray | list[np.ndarray], n_classes: int
-) -> np.ndarray:
-    """
-    Get class weights.
-
-    Based on https://github.com/stardist/stardist/blob/conic-2022/examples/conic-2022/train.ipynb
-
-    Args:
-        Y (np.ndarray): Array of labels (2 channels: instance and class).
-        n_classes (int): Number of classes.
-
-    Returns:
-        np.ndarray: Array of class weights.
-    """
-    class_counts = get_class_counts(
-        np.concatenate([y[1].ravel() for y in Y]), n_classes
-    )
-    inv_freq = np.median(class_counts) / class_counts
-    inv_freq = inv_freq**0.5
-    class_weights = inv_freq.round(4)
-    utils_logger.info(f"class weights = {class_weights.tolist()}")
-    return class_weights
-
-
 def plot_images(images: np.ndarray, labels: np.ndarray | None = None):
     """
     Plot images and labels.
@@ -331,137 +292,6 @@ def plot_images(images: np.ndarray, labels: np.ndarray | None = None):
             ax[i].axis("off")
     plt.tight_layout()
     plt.show()
-
-
-def get_instance_counts(
-    labels: np.ndarray, label_instances: bool = False
-) -> np.ndarray:
-    """
-    Get instance counts.
-
-    Args:
-        labels (np.ndarray): Array of labels (2 channels: instance and class).
-        label_instances (bool, optional): Whether to label instances. Defaults to False.
-
-    Returns:
-        np.ndarray: Array of instance counts.
-    """
-    n_classes = np.max([im[1].max() for im in labels]) + 1
-    counts = np.zeros((len(labels), n_classes))
-    for i in range(len(labels)):
-        if label_instances:
-            instances = label(labels[i][0])
-        else:
-            instances = labels[i][0]
-        classes = labels[i][1]
-        for j in range(n_classes):
-            counts[i, j] = np.unique(instances[classes == j]).size
-    return counts
-
-
-def compute_custom_oversampling_probabilities(
-    labels: np.ndarray, power: float = 1
-) -> np.ndarray:
-    """
-    Compute custom oversampling probabilities using instance-weighted class balancing.
-
-    Args:
-        labels (np.ndarray): Array of labels (2 channels: instance and class).
-        power (float, optional): Power to raise the weights to. Defaults to 1.
-
-    Returns:
-        np.ndarray: Normalized probability array for training sample selection.
-    """
-    classes = np.concatenate([im[1].flatten() for im in labels])
-    n_classes = classes.max() + 1
-    class_counts = get_class_counts(classes, n_classes)
-    class_weights = 1 / class_counts
-    class_weights[0] = 0
-    instance_counts = get_instance_counts(labels)
-    weights = np.sum(instance_counts * class_weights[None], 1)
-    weights = weights**power
-    weights = weights / weights.sum()
-    utils_logger.info(
-        f"Custom oversampling - probability range: {weights.min():.6f} to {weights.max():.6f}"
-    )
-    return weights
-
-
-def compute_stardist_oversampling_probabilities(
-    train_labels: list[np.ndarray] | np.ndarray, n_rare_classes: int = 4
-) -> np.ndarray:
-    """
-    Compute oversampling probabilities using StarDist's methodology for rare class emphasis.
-    Based on https://github.com/stardist/stardist/blob/conic-2022/examples/conic-2022/train.ipynb:
-
-    Args:
-        train_labels: ClassPose labels with shape (N, C, H, W), channel 1 = class labels
-        n_rare_classes: Number of rarest classes to oversample (default: 4)
-
-    Returns:
-        Probability array of shape (N,) normalized to sum to 1.0 for use as train_probs
-    """
-    # Handle both array and list formats
-    if isinstance(train_labels, list):
-        train_labels = np.array(train_labels)
-
-    # Extract class channel - adapt from StarDist's Y0[:, :, :, 1] to ClassPose's [:, 1, :, :]
-    class_channel = train_labels[:, 1, :, :]  # Shape: (N, H, W)
-    n_images = class_channel.shape[0]
-
-    # Count instances per class using 4x downsampling (matches StarDist exactly)
-    # StarDist: np.bincount(Y0[:, ::4, ::4, 1].ravel(), minlength=len(CLASS_NAMES))
-    class_counts = np.bincount(
-        class_channel[:, ::4, ::4].ravel(),
-        minlength=int(class_channel.max()) + 1,
-    )
-
-    # Identify the most infrequent classes (excluding background class 0)
-    rare_classes = np.argsort(class_counts[1:])[:n_rare_classes] + 1
-
-    # Validate that rare classes have non-zero counts
-    for c in rare_classes:
-        if class_counts[c] == 0:
-            utils_logger.warning(f"Rare class {c} has zero instances")
-
-    # Compute scaling factors using square root methodology (matches StarDist)
-    # StarDist: n_extras = np.sqrt(np.sum(class_counts[1:]) / class_counts[extra_classes])
-    total_instances = np.sum(class_counts[1:])  # Exclude background
-    scaling_factors = np.sqrt(total_instances / class_counts[rare_classes])
-
-    utils_logger.info(f"Oversampling rare classes: {rare_classes.tolist()}")
-    utils_logger.info(f"Scaling factors: {scaling_factors.tolist()}")
-
-    # Start with uniform probabilities
-    train_probs = np.ones(n_images, dtype=np.float64)
-
-    # Apply StarDist's probability computation for each rare class
-    for c, scaling_factor in zip(rare_classes, scaling_factors):
-        # Count instances of rare class per image using 2x downsampling (matches StarDist)
-        # StarDist: prob = np.sum(Y0[:, ::2, ::2, 1] == c, axis=(1, 2))
-        prob = np.sum(class_channel[:, ::2, ::2] == c, axis=(1, 2)).astype(
-            np.float64
-        )
-
-        # Apply StarDist's exact methodology
-        prob = np.clip(prob, 0, np.percentile(prob, 99.8))  # Clip outliers
-        prob = prob**2  # Square for emphasis
-
-        # Only update probabilities for images containing this rare class
-        if np.sum(prob) > 0:
-            prob = prob / np.max(prob)
-            # Weight the probabilities by the scaling factor
-            train_probs *= 1.0 + scaling_factor * prob
-
-    # Final normalization to ensure probabilities sum to 1.0
-    train_probs = train_probs / np.sum(train_probs)
-
-    utils_logger.info(
-        f"Probability statistics - min: {train_probs.min():.6f}, "
-        f"max: {train_probs.max():.6f}, median: {np.median(train_probs):.6f}"
-    )
-
-    return train_probs
 
 
 def download_if_unavailable(
